Emacs: switch to MELPA stable for packages
Some packages disabled because they don't have stable versions.  Created
issues on their GH repos and sent an email to maintainer of the other
which is not GH at all.

Also removed the dependence on color-theme because (TIL) it's not
particularly necessary since Emacs 24.  Had to switch to a different
monokai theme which I'm not as happy with, but I can work on reducing
the angry fruit salad quotient.