Fix: CIDER auto-requires wrong user namespace
If project has `foo.bar.user` namespace, then CIDER / CLJR will add:

    (:require [foo.bar.user :as user])

To the namespace.  This is almost always undesirable to me because I
just want the basic `user` namespace.

There appears to be no built-in way to avoid this.  Even adding
`("user" . "user")` to `cljr-magic-require-namespaces` doesn't do the
trick (it leads to a weird error).

This change fixes the issue via advice.